use the same tests for sqlite and postgres
